backends/python/requirements.txt
google.auth
requests